metodos de decisão OK

+++ src/main/java/rest/Service02.java
@@ -81,26 +81,56 @@ public class Service02 {
                 }
             }
             if(flag){
-                //enviar decisao PUT
-                System.out.println("\nDecisão PUT");
-                //realizar a operação
+                for(Replica replica : dataBase.getReplicas()){
+                    WebTarget t = client.target(UriBuilder.fromUri(replica.getEndpoint()).build());
+                    Response response = t.path(replica.getId()).path("decision").request().put(Entity.entity(operation.getId(),MediaType.APPLICATION_JSON), Response.class);
+                    System.out.println("PUT - Status de resposta: "+response.getStatus());
+                }
+                performOperation(operation);
                 this.dataBase.addActions(operation.getId(), "success");
-                return Response.ok().build();
+                return Response.status(201).build();
             }else{
-                //enviar decisao DELETE
-                System.out.println("\nDecisão DELETE");
+                for(Replica replica : dataBase.getReplicas()){
+                    WebTarget t = client.target(UriBuilder.fromUri(replica.getEndpoint()).build());
+                    Response response = t.path(replica.getId()).path("decision").path("/"+operation.getId()).request().delete();
+                    System.out.println("DELETE - Status de resposta: "+response.getStatus());
+                }
+                this.dataBase.removeOperation(operation.getId());
                 this.dataBase.addActions(operation.getId(), "fail");
                 return Response.status(403).build();
             }
         }else{
-            System.out.println("NAO sou coordenador "+name);
             this.dataBase.addOperation(operation);
             float number = dataBase.getRandomNumber();
-            System.out.println("number: "+number);
+            System.out.println("sorteio: "+number);
             if(number <= 0.7){
                 return Response.ok().build();
             }else{
                 return Response.status(403).build();
+            }
+        }
+    }
+
+    private void performOperation(Operation operation){
+        for(Account account : dataBase.getAccounts()){
+            if(account.getNumero().equals(operation.getConta())){
+                String saldo = account.getSaldo();
+                saldo = saldo.replaceAll(",", ".");
+                System.out.println("GET SALDO : " + saldo);
+                double dsaldo = Double.parseDouble(saldo);
+
+                String valor = operation.getValor();
+                valor = valor.replaceAll(",",".");
+                double dvalor = Double.parseDouble(valor);
+
+                if(operation.getOperacao().equals("debito")){
+                    dsaldo = dsaldo - dvalor;
+                }else{
+                    dsaldo = dsaldo + dvalor;
+                }
+                saldo = ""+dsaldo;
+                account.setSaldo(saldo);
+                break;
             }
         }
     }
@@ -112,17 +142,37 @@ public class Service02 {
         if(dataBase.isCoordinator()){
             return Response.status(400).build();
         }
-        this.dataBase.removeOperation(id);
+        boolean flag = false;
+        for(Operation operation : dataBase.getOperations()){
+            if(operation.getId().equals(id)){
+                flag = true;
+                break;
+            }
+        }
+        if(!flag){
+            return Response.status(404).build();
+        }
+        performOperation(this.dataBase.removeOperation(id));
         this.dataBase.addActions(id, "success");
         return Response.ok().build();
     }
 
-    @Path("decision")
+    @Path("decision/{id}")
     @DELETE
-    @Consumes(MediaType.APPLICATION_JSON)
-    public Response rollbackDecision(String id){
+    public Response rollbackDecision(@PathParam("id") String id){
+        //Se ID não existir - retornar o código HTTP 404 Not Found.
         if(dataBase.isCoordinator()){
             return Response.status(400).build();
+        }
+        boolean flag = false;
+        for(Operation operation : dataBase.getOperations()){
+            if(operation.getId().equals(id)){
+                flag = true;
+                break;
+            }
+        }
+        if(!flag){
+            return Response.status(404).build();
         }
         this.dataBase.removeOperation(id);
         this.dataBase.addActions(id, "fail");
@@ -142,20 +192,18 @@ public class Service02 {
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
     public Response loadSeed(int seed) {
-        dataBase.setSeed(seed);
         if (dataBase.isCoordinator()){
             Replica replica = dataBase.getReplicas().get(0);
             WebTarget t = client.target(UriBuilder.fromUri(replica.getEndpoint()).build());
             Response response = t.path(replica.getId()).path("seed").request().post(Entity.entity(12322,MediaType.APPLICATION_JSON), Response.class);
-            System.out.println("Status de resposta: "+response.getStatus());
 
             replica = dataBase.getReplicas().get(1);
             t = client.target(UriBuilder.fromUri(replica.getEndpoint()).build());
             response = t.path(replica.getId()).path("seed").request().post(Entity.entity(12333,MediaType.APPLICATION_JSON), Response.class);
-            System.out.println("Status de resposta: "+response.getStatus());
 
             return Response.ok().build();
         }
+        dataBase.setSeed(seed);
         return Response.ok().build();
     }
